Move duplicate logic into traits
