+++ core/Db/Adapter/Mysqli.php
@@ -21,14 +21,13 @@ use Zend_Db_Adapter_Mysqli;
  */
 class Mysqli extends Zend_Db_Adapter_Mysqli implements AdapterInterface
 {
+    use Db\TransactionalDatabaseDynamicTrait;
+
     /**
      * Constructor
      *
      * @param array|Zend_Config $config database configuration
      */
-
-    private $supportsTransactionLevelForNonLockingReads;
-
     public function __construct($config)
     {
         // Enable LOAD DATA INFILE
@@ -253,29 +252,5 @@ class Mysqli extends Zend_Db_Adapter_Mysqli implements AdapterInterface
         $revision = (int)($version % 100);
 
         return $major . '.' . $minor . '.' . $revision;
-    }
-
-    public function getCurrentTransactionIsolationLevelForSession(): string
-    {
-        try {
-            return $this->fetchOne('SELECT @@TX_ISOLATION');
-        } catch (Exception $e) {
-            return $this->fetchOne('SELECT @@transaction_isolation');
-        }
-    }
-
-    public function setTransactionIsolationLevel(string $level): void
-    {
-        $this->query("SET SESSION TRANSACTION ISOLATION LEVEL $level");
-    }
-
-    public function getSupportsTransactionLevelForNonLockingReads(): ?bool
-    {
-        return $this->supportsTransactionLevelForNonLockingReads;
-    }
-
-    public function setSupportsTransactionLevelForNonLockingReads(bool $supports = null): void
-    {
-        $this->supportsTransactionLevelForNonLockingReads = $supports;
     }
 }
